adicionando interfaces Dao

+++ src/main/java/br/com/padariaweb/dao/IFuncionarioDao.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import br.com.padariaweb.entity.Funcionario;
 
 @Repository
-public interface IFuncionarioDao extends IGenericoCRUDDAO<Funcionario, Integer> {
+public interface IFuncionarioDao extends IGenericoCRUDDAO<Funcionario, Long> {
 
 	public List<Funcionario> pesquisarFuncionario(Funcionario filtro, Integer first, Integer maxPerPage);
